Basic Dockspace
Created a basic dockspace, which allows windows to be moved around. The default positions of these windows need to be set and multi monitor support needs to be added as currently framerate drops drastically and doesn't render properly if main window is on second monitor. Removed unneccesary shaders as fbo is now rendered to ImGui viewport.