Basic Property Panel
Created a basic system for rendering the property oanel for an entity. Components have to implement their own user interface which will be rendered when an entity is selected. The transform component has a basic one which allows the position of an entity to be altered. This UI will be reworked.